refactor(walk): add service and controller packages and redistribut its tests into them
